% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_mGSFPCA.R
\name{eval_mGSFPCA}
\alias{eval_mGSFPCA}
\title{Evaluate Principal Component Functions from mGSFPCA}
\usage{
eval_mGSFPCA(mGSFPCA_obj, eval_pts = NULL, matrix_orth = FALSE)
}
\arguments{
\item{mGSFPCA_obj}{A list object returned by the \code{mGSFPCA} function,
containing the estimated principal components, basis functions, and other
model parameters.}

\item{eval_pts}{Numeric vector specifying the time points at which to
evaluate the principal component functions. The time points should either be
in the original data range or \eqn{\in [0, 1]}, in which case the output is
equivalent to calling fda::eval.fd on mGSFPCA_obj$eigenfunction.}

\item{matrix_orth}{Logical indicating whether to enforce strict orthogonality
of the evaluated eigenfunctions \eqn{\Phi^\top \Phi = I_p} under standard
matrix multiplication (discrete inner product). Diagnostic comparison plots
are displayed. Requires \code{length(eval_pts) > p}.}
}
\value{
A matrix of dimension \code{length(eval_pts) x p}, where \code{p} is
the number of principal components, containing the evaluated principal
component functions at the specified \code{eval_pts}.
}
\description{
Evaluates the principal component functions (eigenfunctions) from mGSFPCA()
object at specified time points.
}
\details{
The \code{eval_mGSFPCA} function evaluates the principal component functions
(eigenfunctions) from an \code{mGSFPCA} object at the provided \code{eval_pts}.
}
