% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_spMultFPCA.R
\name{eval_spMultFPCA}
\alias{eval_spMultFPCA}
\title{Evaluate Principal Component Functions from spMultFPCA}
\usage{
eval_spMultFPCA(spMultFPCA_obj, eval_pts = NULL)
}
\arguments{
\item{spMultFPCA_obj}{A list object returned by the \code{spMultFPCA}
function, containing the estimated principal components, basis functions,
and other model parameters.}

\item{eval_pts}{Numeric vector specifying the time points at which to
evaluate the principal component functions.}
}
\value{
A list containing the following components:
\itemize{
\item \code{Ckk}: A list containing the eigenfunctions, scores, and curve
prediction for each variable at eval_pts.
\item \code{scores}: A matrix of multivariate principal component scores
for each subject.
\item \code{eigFunctions}: A list of matrices, each containing the
multivariate eigenfunctions for a variable.
\item \code{values}: A vector of multivariate eigenvalues.
\item \code{vectors}: A matrix representing the eigenvectors associated
with the combined univariate score vectors
\item \code{npc}: The number of multivariate principal components retained.
\item \code{Cov}: The estimated multivariate covariance matrix.
\item \code{fullEigFun}: A matrix of concatenated multivariate
eigenfunctions across all variables.
}
}
\description{
Evaluates the principal component functions (eigenfunctions) from
spMultFPCA() object at specified time points (eval_pts).
}
\details{
The \code{eval_spMultFPCA} function evaluates the principal component
functions (eigenfunctions) from an \code{spMultFPCA} object at the provided
\code{eval_pts}.
}
