% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xHat.R
\name{get_xHat}
\alias{get_xHat}
\title{Compute Fitted Trajectories and Scores from mGSFPCA}
\usage{
get_xHat(mGSFPCA_obj, eval_pts = NULL, alpha = 0.05)
}
\arguments{
\item{mGSFPCA_obj}{A list object returned by the \code{mGSFPCA} function,
containing the estimated principal components, mean function, and other model
parameters.}

\item{eval_pts}{Numeric vector specifying the time points at which to
evaluate fitted trajectories. If \code{NULL} (default), the evaluation grid
from the \code{mGSFPCA_obj$pars$evalGrid} is used.}

\item{alpha}{Numeric value between 0 and 1 specifying the significance level
for constructing point-wise confidence bands for the predicted trajectories.
Default is 0.05.}
}
\value{
A list containing the following components:
\itemize{
\item \code{xHat}: A matrix of dimension \code{n x length(eval_pts)}, where
\code{n} is the number of subjects, containing the fitted trajectories for
each subject evaluated at \code{eval_pts}.
\item \code{Xi}: A matrix of dimension \code{n x p}, where \code{p} is the
number of principal components, containing the principal component scores
for each subject.
\item \code{xHat_CI}: A matrix of dimension \code{n x length(eval_pts)},
containing the (1-\eqn{\alpha}) point-wise confidence intervals for each
subjects at  \code{eval_pts}.
}
}
\description{
Estimates principal component scores and predicts trajectories for each
subject using the results from mGSFPCA().
}
\details{
The \code{get_xHat} function computes the fitted trajectories and principal
component scores for each subject based on the output of the \code{mGSFPCA}
function. It uses the PACE (Principal Analysis by Conditional Expectation)
method to estimate the scores.
}
\references{
Yao, F., Müller, H.-G., & Wang, J.-L. (2005). Functional data analysis for
sparse longitudinal data. Journal of the American Statistical Association,
100(470), 577–590.
}
